dashboard: show outdated loans. Use a due-date, not max-time.
The due-date was accepted to be null, otherwise, when migrating the
tables, Mito complains that the non-nil constraint is not
fullfilled. We should be able to provide a default, or to run data
migrations. See https://github.com/fukamachi/mito/issues/109

Data was migrated manually:

(loop for cc in *CONTACT-COPIES*
             do (setf (due-date cc)
                      (print (local-time:timestamp+ (mito:object-created-at cc)
                                             (max-time cc)
                                             :day)))
             (mito:save-dao cc))

Mito was updated from 2021-04 (buggy, would not save the object) to
2021-08 (better).