Card: boost the search by shelf with HTMX
The <a/> link is followed when loading the card page, the SQL query is
run on the server asynchronsouly and the new HTML page is stored by HTMX,
waiting for us. If we click the <a/> link finally, the page can be
displaed instantly.

It really feels snappier: the page change is done in 16ms.
What saves us a ton of time is that, because there is no page reload,
the browser doesn't re-load the JS and CSS dependencies.

[ci skip]