Use a feature flag to compile templates in memory
The setting is set in the Makefile, for the standalone binary only.

That way, respect the development settings when we quickload the app:
we want a file-system store and we want to reload the templates when
we reload the page.

That being said, there might be an issue, when I quickload the app I
see a MEMORY store is created. After a C-c C-c, we are good.

fixes #33